pyproject.toml
[project]
name = "travel-planner-agent"
version = "0.1.0"
description = "旅游行程策划智能助手 - 基于 LangChain 1.0+"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-milvus>=0.2.0",
    "langsmith>=0.4.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "sse-starlette>=2.1.3",
    "deepagents>=0.3.8",
    "dashscope>=1.0.0",
]

[project.scripts]
travel-agent = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
